+++ packages/dumb/src/index.ts
@@ -54,15 +54,14 @@ import {
 } from './components/stats/average-daily-dose/average-daily-dose-stat'
 import { SimpleStatMemoized as SimpleStat } from './components/stats/simple/simple-stat'
 import { LabeledCheckbox } from './components/controls/labeled-checkbox/labeled-checkbox'
-import { CbgMedianAnimated } from './components/trends/cbg/cbg-median-animated/cbg-median-animated'
 import { Table } from './components/settings/table'
 import CbgSliceSegment from './components/trends/cbg/cbg-slice/cbg-slice-segment'
 import { NoDataLabel } from './components/trends/common/no-data-label/no-data-label'
-import { CbgSliceAnimated } from './components/trends/cbg/cbg-slice/cbg-slice-animated'
 import { BgPrefs } from './models/blood-glucose.model'
 import { XAxisLabels } from './components/trends/common/x-axis-labels/x-axis-labels'
 import { XAxisTicks } from './components/trends/common/x-axis-ticks/x-axis-ticks'
 import { CbgDateTraceLabel } from './components/trends/cbg/cbg-date-trace/cbg-date-trace-label'
+import { CbgSlicesContainerMemoized as CbgSlicesContainer } from './components/trends/cbg/cbg-slice/cbg-slices-container'
 
 export {
   AverageDailyDoseStat,
@@ -72,9 +71,8 @@ export {
   BolusTooltip,
   CbgDateTraceLabel,
   CBGMeanStat,
-  CbgMedianAnimated,
   CBGPercentageBarChart,
-  CbgSliceAnimated,
+  CbgSlicesContainer,
   CbgSliceSegment,
   CBGStandardDeviation,
   CBGStatType,